Code review for Tianjing's CSS file
Tianjing developed a lot of cool ways to style the html. I have never seen @media or @keyframes before reviewing this. This is something that I need to learn and implement on my project so that it can be responsive on any devices. Overall, I am very inspired and learnt a great deal again.